pyproject.toml
[project]
name = "fastmcptest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.8.0",
    "fastmcp>=3.0.0b1",
    "motor>=3.7.1",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.1",
]
